requirements.txt
Click==7.0
Flask==1.0.2
Flask-JWT-Simple==0.0.3
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.1.0
psycopg2-binary==2.7.6.1
PyJWT==1.7.1
Werkzeug==0.14.1
